fix txn_heart_beat test
