+++ src/storage/txn/actions/tests.rs
@@ -2,19 +2,27 @@
 
 //! This file contains tests and testing tools which affects multiple actions
 
+use std::sync::Arc;
+
 use concurrency_manager::ConcurrencyManager;
 use kvproto::kvrpcpb::{
-    Assertion, AssertionLevel, Context,
+    Assertion, AssertionLevel, Context, ExtraOp,
     PrewriteRequestPessimisticAction::{self, *},
 };
 use prewrite::{prewrite, CommitKind, TransactionKind, TransactionProperties};
-use tikv_kv::SnapContext;
+use tikv_kv::{SnapContext, Statistics};
 
 use super::*;
 use crate::storage::{
     kv::WriteData,
+    lock_manager::MockLockManager,
     mvcc::{tests::write, Error, Key, Mutation, MvccTxn, SnapshotReader, TimeStamp},
-    txn, Engine,
+    txn,
+    txn::{
+        commands::{Flush, WriteContext, WriteResult},
+        txn_status_cache::TxnStatusCache,
+    },
+    Engine,
 };
 
 pub fn must_prewrite_put_impl<E: Engine>(
@@ -197,6 +205,87 @@ pub fn must_prewrite_put<E: Engine>(
         Assertion::None,
         AssertionLevel::Off,
     );
+}
+
+pub fn flush_put_impl<E: Engine>(
+    engine: &mut E,
+    key: &[u8],
+    value: impl Into<Vec<u8>>,
+    pk: impl Into<Vec<u8>>,
+    start_ts: impl Into<TimeStamp>,
+    generation: u64,
+    should_not_exist: bool,
+) -> txn::Result<WriteResult> {
+    flush_put_impl_with_assertion(
+        engine,
+        key,
+        value,
+        pk,
+        start_ts,
+        generation,
+        should_not_exist,
+        Assertion::None,
+    )
+}
+
+pub fn flush_put_impl_with_assertion<E: Engine>(
+    engine: &mut E,
+    key: &[u8],
+    value: impl Into<Vec<u8>>,
+    pk: impl Into<Vec<u8>>,
+    start_ts: impl Into<TimeStamp>,
+    generation: u64,
+    should_not_exist: bool,
+    assertion: Assertion,
+) -> txn::Result<WriteResult> {
+    let key = Key::from_raw(key);
+    let start_ts = start_ts.into();
+    let mut m = if should_not_exist {
+        Mutation::make_insert(key, value.into())
+    } else {
+        Mutation::make_put(key, value.into())
+    };
+    m.set_assertion(assertion);
+    let cmd = Flush::new(
+        start_ts,
+        pk.into(),
+        vec![m],
+        generation,
+        3000,
+        AssertionLevel::Strict,
+        Context::new(),
+    );
+    let mut statistics = Statistics::default();
+    let cm = ConcurrencyManager::new(start_ts);
+    let context = WriteContext {
+        lock_mgr: &MockLockManager::new(),
+        concurrency_manager: cm.clone(),
+        extra_op: ExtraOp::Noop,
+        statistics: &mut statistics,
+        async_apply_prewrite: false,
+        raw_ext: None,
+        txn_status_cache: Arc::new(TxnStatusCache::new_for_test()),
+    };
+    let snapshot = engine.snapshot(Default::default()).unwrap();
+    cmd.cmd.process_write(snapshot.clone(), context)
+}
+
+pub fn must_flush_put<E: Engine>(
+    engine: &mut E,
+    key: &[u8],
+    value: impl Into<Vec<u8>>,
+    pk: impl Into<Vec<u8>>,
+    start_ts: impl Into<TimeStamp>,
+    generation: u64,
+) {
+    let res = flush_put_impl(engine, key, value, pk, start_ts, generation, false);
+    assert!(res.is_ok());
+    let res = res.unwrap();
+    let to_be_write = res.to_be_write;
+    if to_be_write.modifies.is_empty() {
+        return;
+    }
+    engine.write(&Context::new(), to_be_write).unwrap();
 }
 
 pub fn must_prewrite_put_on_region<E: Engine>(
